+++ src/main/java/frc/robot/GlobalConstants.java
@@ -4,7 +4,12 @@ public class GlobalConstants {
     // Comp
     public static final Boolean COMP = false;
     
-    // SysID tuning
+    // System Identification
     public static final Boolean SYS_ID = false;
     public static final String CTRE_LOG_PATH = "/media/sda1/ctre-logs/";
+
+    // Driver Preferences
+    public static final double VEL_CONTROL_BASE = 0.2;
+    public static final double DEADBAND_TRANSLATION = 0.15;
+    public static final double DEADBAND_ROTATION = 0.15;
 }